binance utils refactor
