+++ src/base/utils/utils.js
@@ -37,7 +37,11 @@ const timeIntervalToSeconds = (interval) => {
     return seconds;
 };
 
-const createHmacSignatureDefault = ({ ts, method, path, body = '' }, key, encoding) => {
+const createHmacSignatureDefault = (
+    { ts, method, path, body = '' },
+    key,
+    encoding,
+) => {
     if (body) body = JSON.stringify(body);
 
     const signaturePayload = [ts, method, path, body].join('');
